refactor audit events 1st round
